+++ BudgetApplication/app/src/main/java/com/example/aaron/welcomeActivity/expense.java
@@ -8,24 +8,24 @@ import java.io.Serializable;
 public class Expense implements Serializable {
     private String name; //name of Expense
     private long budgetID = 0;
-    private float currentValue = 0; //how much is currently being spent on the item
-    private float maxValue = 100; //max amount of money that can be spent on item
+    private double currentValue = 0; //how much is currently being spent on the item
+    private double maxValue = 100; //max amount of money that can be spent on item
     private long IDNumber = 0; //ID number that corresponds to the SQL database location
     private int priority = 0; //Priority that corresponds to the importance of the Expense
     private int aisle = 0; //What aisle the product is in at the store you frequent. Zero = no aisle
     private String paymentInterval = null;
 
-    Expense(String newName, float newCurrent, float newMax){
+    Expense(String newName, Double newCurrent, Double newMax){
         name = newName;
         currentValue = newCurrent;
         maxValue = newMax;
     }
 
-    void setCurrentExpense(float newCE){
+    void setCurrentExpense(Double newCE){
         currentValue = newCE;
     }
 
-    void setMaxExpense(float newME){
+    void setMaxExpense(Double newME){
         maxValue = newME;
     }
 
@@ -43,11 +43,11 @@ public class Expense implements Serializable {
 
     void setBudgetID(long newID) { budgetID = newID; }
 
-    float getCurrentExpense(){
+    double getCurrentExpense(){
         return currentValue;
     }
 
-    float getMaxExpense(){
+    double getMaxExpense(){
         return maxValue;
     }
